refactor (shell) : Prerequisite code reorganization for gopsutil windows refactor
Related to #4588

Move shell detection related methods from shell_unix.go to shell.go so
that they can be accessed from Windows code as well
