--- pkg/os/shell/shell_linux.go
@@ -1,5 +1,0 @@
-package shell
-
-var (
-	supportedShell = []string{"bash", "zsh", "fish"}
-)
